fix external XWindow resize
If plotting to an XWindow and the window was resized by the user before
re-plotting, the draw surface would not get updated to reflect the new
XWindow size.

Fixes necessary:
- detect if the XWindow was resized (in handle change to new page event) and set resize flag for the device in such a case
- the PANEL size(s) were not updated: this was the main culprit.

The window and cairo surface were always changed upon a new XWindow page but
the panel sizes, kept internally in units of pixels(!) were never updated.
Thus even if the XWindow was made smaller/larger, the amount of pixels
actually drawn was never changed.