Support for backspace (\b) in printing strings
For time labelling to look good it is 'necessary' to make the superscript
appear on top of e.g. the decimal point. This can be done if 'backspace' is
supported - backing up the text by one or more characters.

This necessitated rewriting the giza_parse_string() a bit because in the
current version backing up to a previous character's position was
impossible.