Support for PGBOX 'I' and 'P' options and bug fix
giza's pgbox(...) routine did not support the PGPLOT options 'I' and 'P':
 -I: Invert tick marks - write them outside the viewport in stead of inside
     (default)
 -P: Project major tick marks outside the viewport (ignored if 'I' also
     specified)

Fixed bug where tickmarks on 'the other side' of the frame were drawn in the
same direction but they need to be inverted:

If tick marks on the bottom are drawn inwards (i.e. up) then at the top they
must be drawn downwards to achieve the same effect. The existing code drew
the top tick marks outside the viewport because it would draw them 'up' as
well.