Add support for 'M' option to PG(T)BOX
When it comes to writing labels PG(T)BOX supports the following:

  -N :- Label the axis (conventional, below/left viewport)
  -M :- Put labels in the unconvential location (above/right viewport)

The 'M' option was not implemented by giza so I added it.
Whilst at it removed a few levels of somewhat unnecessary if()'s by adding a
high-level check and sometimes continue'ing from a loop if the rest of the
loop body is precluded from being executed anyway.