Fixed earlier broken 'repair'
Sometimes the PGPLOT doc is somewhat ambiguous but not for 'N' and 'M'
options: both imply labels have to be drawn, only 'M' sais to do so in the
unconvential location.

Ambiguity example: does 'L' (logarithmic labelling) *imply* 'N' or 'M'
("write labels")? Or is 'L' only honoured if either 'N' or 'M' is specified
(which is how the code currently works).