Better XWindow resizing implementation
It turned out that the XWindow resizing wasn't handled correctly in all
cases; notable pgdemo13 (with two x-windows) exposed a problem in the code in the first
window, the one with the choice menu.

More often than not it would show ellipses where circles should be drawn and
the text was (way) ouside the window.

What happened was two calls to giza_change_page_xw() quickly after each other. The first had Dev[id].resize = 1 (from PGPAP(...) in pgdemo13), which triggered an XResizeWindow(...) in the code, to change the size of the XWindow to the requested device size. The next call entered with Dev[id].resize = 0 (because that was already handled) ... but that one detected a window geometry change! Upon inspection it turned out that, as far as X11 was concerned, the window still had the /old/ size whilst our internal bookkeeping was already updated with the requested size.

Turns out that the XResizeWindow() does not take effect immediately. So the code now waits for ConfigureNotify events until the requested window size has been achieved. Now it works reliably and windows can be resized between plotting.