Rework giza_parse_string, support Hershey and more
Looking at the pgdemo2 page with all the font, symbol and marker support,
there were a few things missing from Giza, the biggest by far is lack of
most of the Hershey symbols "\(nnnn)". Some three symbols were in the lexer
and had their own token.

Hershey has ~4000 symbols so inventing ~4000 tokens for each Hershy symbol
individually seemed ... counterproductive. A similar issue was with the
greek/Greek alphabet, each letter being their own token and having also an
alias (TeX "\<letter>" syntax) and some missing PGPLOT escape sequences like
"\A" for the Angstrom symbol and "\." for multiplication dot. There was also
support for TeX style font changes "\rm" &cet.

A.) The new giza_parse_string(...) is not recursive anymore but handles nested
super/sub script and font changes by keeping a stack of those states.

B.) To support as many Hershey symbols as possible, started from
    https://github.com/reinholtz/hershey-fonts-with-unicode
to fill the lookup table of number to UTF strings that were already defined
using 'hersheyunicodemap.csv'. Updated that table by looking up the unicode
name from 'hersheyindex.txt' in
    http://www.unicode.org/Public/UNIDATA/UnicodeData.txt
and substitute "\\ufffd" entries (unicode replacement character) with a match.
Cyryllic characters now show up Just Fine (tm) in the pgdemo2 screenshot.

C.) The lexer is now much symplified and is enriched with validation such that
the code handling the tokens can be simplified to do so without checking.
Translating giza-scanner.l into lex.yy.c does not yield a warning anymore.

D.) To handle all "\<...>" sequences easily, defined a token remapper table. If
the lexer encounters a "\blah" construct it looks up if "blah" is in the
remap table. If so, the string parser will remap the "\blah" sequence to
another token with another token text.

Examples:
        /* "\Sun" remapped to Hershey symbol 2281 */
        {"Sun",      {GIZA_TOKEN_HERSHEY, "2281"   }},
        /* "\rm" is an alias for \fr */
        {"rm" ,      {GIZA_TOKEN_FONT   , "r"      }},
        /* etc ... */
        {"Lambda",   {GIZA_TOKEN_GREEK  , "L"      }},

Not all utf strings show up in the screenshot but that has to do with the
selection of fonts available; the used fonts do not have glyphs for all the
unicode code points. The "cursive" font has most of them, but also not all.

E.) The new giza_parse_string(...) implements backspace to the extent that:
                x^2\b_i
renders as x-i-squared proper with the super+sub scripts where they should.
This was necessary for PGTBOX to have the unit of the time superscripted on
top of the decimal point. I added an extra test case to pgdemo2.f to
illustrate.

F.) The markers \mnn were drawn very small so made them a little larger, now
they look better (in the pgdemo, that is)