Fix backspace handling ...
The meaning of the nGlyph index used in giza_parse_string changed subtly
during development and the handling of backspace was not changed
accordingly.

The result was that "x^2\b_i" rendered fine but the unit superscript over
the decimal point in PGTBOX() now looked horrible because the pen position
was moved to an erroneous position. Sigh.

In order to make "x-i-squared" render nicely it must now be spelled as:
    x^2\b\b_i
with two backspaces. The first backs up over the superscript '2', the second
backs up over the '^' superscript operation; when entering the superscript
the baseline of the font and the current pen position is changed
immediately. So if that is not undone before staring to render further
characters, the pen position is not where you hoped it would be.