Support cpgqinf(), fix critical bug
Found out that the fortron giza pgplot binding implemented PQQINF() (to
query version &cet) but the C-language binding didn't.

Whilst implementing cpgqinf() found that giza_query_device() and
_giza_int_to_device() contained a serious bug, with SIGSEGV looming or loss
of information.

Both quoted functions copy information into a user allocated string buffer.
Extract of relevant code with 'noise' deleted:

    _giza_int_to_device (int numDevice, char *DeviceName)
    {
        case GIZA_DEVICE_NULL:
          strncpy(DeviceName, "/null", sizeof(DeviceName)-1);
          break;
    ....

The user allocated buffers are passed by pointer ("char* DeviceName") and
the 'sizeof(...)' operator returns the size of the _pointer_ not the size of
the buffer it's pointing at!

Depending on the platform "sizeof(char*) - 1" evaluates to 3 (32 bit) or 7
(64 bit), irrespective of how many characters the user really allocated.

This can lead to either:

    - overestimating the amount of characters available in case the user
      allocated less than 3 (or 7) bytes for their buffer. A strncpy() with
      a source longer than 3 (or 7) would then overwrite memory that it
      shouldn't, which in turn would lead to unpredictable behaviour at best
      and program termination via SIGSEGV at worst.

    - underestimating the amount characters available if the user allocated
      more than 3 (or 7) characters for their target buffer. In this case,
      the strncpy() with a source string > 3 (or 7) characters would
      truncate the output - e.g. the file name of a device would easily
      be > 7 characters long but at most 3 (or 7) of them would be copied.
      This would lead to unnecessary loss of information.

In order to fix this the target buffer size must be explicitly passed to the
functions by the caller. Thus the prototype of giza_query_device() needed to
be altered and as a result the Fortran wrapper needed adapting to the new
signature and pass in the extra variable accordingly.