`giza_format_number()` string buffer length
The `giza_format_number()` function writes unconditionally to a destination
string buffer without being informed how large that buffer is.

This commit adds an extra parameter to be passed to `giza_format_number()` -
the size of the string buffer. It changes the implementation to take care of
not writing outside the destination string's boundaries.

It should be noted that this may have an effect on cpgnumb()/PGNUMB()
because there is no officially documented necessity that the *string_length
parameter in cpgnumb() has to contain the size of the buffer upon entry; the
documentation only mentions the contents of that parameter are set on output
to the number of characters written.

I have checked the official PGPLOT libcpgplot binding and perl5-PGPLOT which
both do this: both write the size of the FORTRAN string in the parameter
before calling cpgnumb() or use a temporary variable and pass that size in.
I have updated giza-fortran.F90 to do the same. The Python-PGPLOT binding
(ppgplot) does not expose cpgnumb() so at this point that's not an issue.