Fix device specific close fn's
giza_open() now detects and handles errors occuring during setup of the new
device. The device specific _giza_close_<device>() functions needed to be
updated to deal with partially/not fully initialized devices.

For the XWindow driver, its shadow administration of GIZA_XWindow devices
(the "XW[...]" array) needed to be managed slightly different, not unlinke
how devices are now initialized/de-initialized in giza-drivers.c

The global 'xid' variable was removed and replaced by a lookup in the
(already existing) giza_xw_id[...] table.

Thinking about it, it seems to me the whole shadow 'xid' administration can
be simplified if Dev[N], if Dev[N] refers to an XWindow, stores its
peculiars in XW[N]. `giza_open()` now already takes care of finding an empty
slot and there should be no reason for the XWindow driver to use that
assigned slot as well.

A separate commit implementing that will follow, which makes that change
opt-in.