Synchronize XW device id with giza_device_t
The global variable 'id' is now also used to address the instance of the
GIZA_XWindow struct where giza_device_t Dev[id]'s properties are stored.

This can be done because:
- giza-driver-xw.c implements a shadow array, "XW",
of euqal length as the Dev[...] array (GIZA_MAX_DEVICES
- `giza_open_...()` now searches for an unused entry in Dev[...]
when opening a device
- `giza_close_device()` closes the correct device when it's, well,
closed. If the entry happened to be an XWindow, the corresponding XW[id]
struct will now also be properly closed.