Fix giza_set_colour_table to set colourindices
The function would warn about inconsistent entries in the controlPoints and
return with an error, not setting the colour-index table.

PGPLOT pgdemo4 (demonstrating PGIMAG) has a table with controlpoints outside
the range 0->1 (-.5, 1.7) triggering this warning and error exit.

The colour table is installed nonetheless!
Net effect: pixel colourings in "true colour" - i.e. those that do not go
through colour indices - keep on working, but those relying on colour
indices don't.

This fix prints the warning, checks that a sensible number of control
points remain, and then goes on computing the colours for the set colour
index range (PGSCIR) - as per documentation.